requirements.txt
tensorflow-gpu==2.0.0
numpy==1.18.1
pandas==1.0.1
scikit-learn==0.22.1